Complete step 1
Where we created the application using the Angular CLI and
performed simple modification to the initial APP component
of our page.

Note, that the component declaration allows us to specify where
the compiler is to source the HTML representation of the "component"
as well as its style. This is much better than the approach I have
seen where all this code is 'embedded' in the component.

    --- Lionel Saliou, Ph.D