Split Master and Detail view into components
Performed simple refactoring to move the detail view away from
the Master view. This will allow for each to evolve independently
from one another other.

I would have prefered for the dependency to be outlined in the
"main HTML" page. Here, whilst it is true that both can evolve
seperately, the 'master' view still requires the 'detail' view
to be present.

    --- Lionel Saliou, Ph.D