Implementing Angular routing
In terms of approach, this is very similar to the way Django does
it. Indeed, we define URLs and point to specific view(s) [in
this case component templates].

The styling is bit odd at the moment because of the use of links
but it is likely to be solved shortly with better CSS.

As expected, we finally split the master and detail view completely.
Now, the master view does not know about this view at all. Now,
it is the responsibily of the "routing" to figure out what to
do when the user click on links.

   -- Lionel Saliou, Ph.D