Application completed - search enabled.
This is the final commit for this tutorial.

There is now a search box and a style result set.

Had to remove the application and the search from the
centred div in the APP HTML page.

It was seriously screwing with the layout of the list
and the forms.

I must say it is neat to see the messages added to the application
but it is rather distracting as the user clicks around the application.

   --- Lionel Saliou, Ph.D